separate cloud client implementation
